+++ app/ui/legacy/src/main/java/com/fsck/k9/ui/messagesource/MessageHeadersFragment.kt
@@ -10,29 +10,19 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
-import com.fsck.k9.DI
 import com.fsck.k9.controller.MessageReference
 import com.fsck.k9.mail.Header
 import com.fsck.k9.mail.internet.MimeUtility
 import com.fsck.k9.mailstore.MessageRepository
 import com.fsck.k9.ui.R
+import com.fsck.k9.ui.withArguments
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
+import org.koin.android.ext.android.inject
 
 class MessageHeadersFragment : Fragment() {
-    companion object {
-        private const val ARG_REFERENCE = "reference"
-        fun newInstance(reference: MessageReference): MessageHeadersFragment {
-            val fragment = MessageHeadersFragment()
-            val args = Bundle()
-            args.putString(ARG_REFERENCE, reference.toIdentityString())
-            fragment.arguments = args
-            return fragment
-        }
-    }
-
-    private var messageRepository = DI.get(MessageRepository::class.java)
+    private val messageRepository: MessageRepository by inject()
     private var sourceText: TextView? = null
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
@@ -72,5 +62,15 @@ class MessageHeadersFragment : Fragment() {
             sb.append(MimeUtility.unfoldAndDecode(value))
         }
         sourceText!!.text = sb
+    }
+
+    companion object {
+        private const val ARG_REFERENCE = "reference"
+
+        fun newInstance(reference: MessageReference): MessageHeadersFragment {
+            return MessageHeadersFragment().withArguments(
+                ARG_REFERENCE to reference.toIdentityString()
+            )
+        }
     }
 }
